refactor(gridsat): rename netcdf_metadata module to metadata
