+++ src/goesdl/gridsat/metadata.py
